refactor(server): Move RulesManager to zally-core #560
