+++ server/src/main/java/de/zalando/zally/configuration/RulesManagerConfiguration.kt
@@ -1,7 +1,7 @@
 package de.zalando.zally.configuration
 
 import com.typesafe.config.Config
-import de.zalando.zally.rule.RulesManager
+import de.zalando.zally.core.RulesManager
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
