+++ server/src/main/java/de/zalando/zally/configuration/RulesValidatorConfiguration.kt
@@ -5,7 +5,7 @@ import de.zalando.zally.rule.ApiValidator
 import de.zalando.zally.rule.CompositeRulesValidator
 import de.zalando.zally.rule.ContextRulesValidator
 import de.zalando.zally.rule.JsonRulesValidator
-import de.zalando.zally.rule.RulesManager
+import de.zalando.zally.core.RulesManager
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
