+++ server/src/main/java/de/zalando/zally/rule/ContextRulesValidator.kt
@@ -3,6 +3,7 @@ package de.zalando.zally.rule
 import com.fasterxml.jackson.core.JsonPointer
 import de.zalando.zally.core.ContentParseResult
 import de.zalando.zally.core.DefaultContextFactory
+import de.zalando.zally.core.RulesManager
 import de.zalando.zally.rule.api.Context
 
 /**
