+++ server/src/main/java/de/zalando/zally/rule/RulesValidator.kt
@@ -1,6 +1,7 @@
 package de.zalando.zally.rule
 
 import com.fasterxml.jackson.core.JsonPointer
+import de.zalando.zally.core.CheckDetails
 import de.zalando.zally.core.ContentParseResult
 import de.zalando.zally.core.ObjectTreeReader
 import de.zalando.zally.core.ContentParseResult.NotApplicable
@@ -9,6 +10,7 @@ import de.zalando.zally.core.ContentParseResult.ParsedWithErrors
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
 import de.zalando.zally.core.JsonPointerLocator
+import de.zalando.zally.core.RulesManager
 import de.zalando.zally.core.RulesPolicy
 import org.slf4j.LoggerFactory
 import java.lang.reflect.InvocationTargetException
