+++ server/src/test/java/de/zalando/zally/rule/RulesValidatorTest.kt
@@ -6,6 +6,7 @@ import com.typesafe.config.ConfigFactory
 import de.zalando.zally.core.AbstractRuleSet
 import de.zalando.zally.core.ContentParseResult
 import de.zalando.zally.core.EMPTY_JSON_POINTER
+import de.zalando.zally.core.RulesManager
 import de.zalando.zally.core.RulesPolicy
 import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Rule
