+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/RuleDetails.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Rule
