+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/RulesManager.kt
@@ -1,7 +1,6 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import com.typesafe.config.Config
-import de.zalando.zally.core.RulesPolicy
 import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.RuleSet
